+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [6.18.2] 2024-11-17
 
 - Update README badges labels
+- Upgrade dependencies
 
 ## [6.18.1] 2024-11-16